update with role

+++ Services/CharacterService/CharacterService.cs
@@ -1,12 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Learningcsharp.Controllers.Dtos.Characters;
 using Learningcsharp.Data;
 using Learningcsharp.Dtos.Characters;
 using Learningcsharp.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace Learningcsharp.Controllers.Services.CharacterService
@@ -23,9 +25,13 @@ namespace Learningcsharp.Controllers.Services.CharacterService
         };
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _context = context;
             _mapper = mapper;
 
@@ -35,9 +41,11 @@ namespace Learningcsharp.Controllers.Services.CharacterService
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.Users = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+            serviceResponse.Data = (_context.Characters.Where(c => c.Users.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
@@ -47,10 +55,23 @@ namespace Learningcsharp.Controllers.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(item => item.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                Character character = await _context.Characters
+                        .FirstOrDefaultAsync(item => item.Id == id && item.Users.Id == GetUserId());
+
+                if(character != null) 
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = (_context.Characters.Where(c => c.Users.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "No Persmission to delete this character";
+                }
+                
+                
             }
             catch (Exception ex)
             {
@@ -64,7 +85,8 @@ namespace Learningcsharp.Controllers.Services.CharacterService
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.Characters.ToListAsync();
+            // List<Character> dbCharacters = await _context.Characters.ToListAsync();
+            List<Character> dbCharacters = await _context.Characters.Where(c => c.Users.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
@@ -72,7 +94,7 @@ namespace Learningcsharp.Controllers.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(num => num.Id == id);
+            Character dbCharacter = await _context.Characters.Include(c => c.Users ).FirstOrDefaultAsync(num => num.Id == id && num.Users.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -81,26 +103,36 @@ namespace Learningcsharp.Controllers.Services.CharacterService
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
 
-            try {
-            Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
-            character.Name = updateCharacter.Name;
-            character.Class = updateCharacter.Class;
-            character.Defense = updateCharacter.Defense;
-            character.HitPoints = updateCharacter.HitPoints;
-            character.Intelligence = updateCharacter.Intelligence;
-            character.Strength = updateCharacter.Strength;
+            try
+            {
+                Character character = await _context.Characters.Include(c => c.Users).FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
+                if(character.Users.Id == GetUserId())
+                {
+                character.Name = updateCharacter.Name;
+                character.Class = updateCharacter.Class;
+                character.Defense = updateCharacter.Defense;
+                character.HitPoints = updateCharacter.HitPoints;
+                character.Intelligence = updateCharacter.Intelligence;
+                character.Strength = updateCharacter.Strength;
 
-            _context.Characters.Update(character);
-            await _context.SaveChangesAsync();
+                _context.Characters.Update(character);
+                await _context.SaveChangesAsync();
 
-            serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                } 
+                else 
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Sorry you can't edit this character";
+                }
+                
             }
             catch (Exception ex)
             {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
-            
+
             return serviceResponse;
 
         }
